refactor: update import paths for asset_helper module
