+++ services/asset_helper.py
